Add files via upload
This is a chatbot that I trained and wrote with a very simple interface. It can be run locally. If you want to access the chat interface within the LAN, you need to do cross-domain and network penetration. I have tried it once before and it was successful. Unfortunately Unfortunately, I forgot to record my operation... This is a chatbot that I trained and wrote with a very simple interface. It can be run locally. If you want to access the chat interface within the LAN, you need to do cross-domain and network penetration. I have tried it once before and it was successful. Unfortunately Unfortunately, I forgot to record my operation process, so if you want to try it yourself, please find the solution yourself. If you succeed, please contact me and I can add instructions.